KAgent CLI Fixes (#762)
# Summary

Running through the available commands and seeing what can be fixed or
improved.

## Non-Interactive

- [x] bug_report (works)
- Added `memory`, `mcpserver`, and `remotemcpserver` to bug report. Can
revert if not necessary.
- [x] get
    - Fixed.
- The issue with these commands was due to the setup we had with the Go
client. The CLI would set `apiUrl=kagent-svc/api`, and the _client_
would do request through a command using `/api/<target>` as the path.
This meant the final request to the API would be:
`{client_url}{command_path} -> kagent-svc/api/api/<target>`. Duped
`/api`.
- The fix I implemented was to move our CLI from using `apiUrl` to
`kagentUrl`. It's simply the URL without the `/api` portion. This keeps
the Go client acting as it should (the middleware(?) to *all* controller
paths, API and non-API.
    - Two other ways to do it would be to:
        1. Update the client methods to remove the `/api`.
1. I decided against this because the client exposes some non `/api`
endpoints. So if we keep passing `apiUrl` from the CLI, then if we do a
non-api path (e.g. `/health`), we'd accidentally do
`kagent-svc/api/health` instead of `kagent-svc/health`.
2. Pass in `apiUrl` to the client, and do a strip of `/api` post-fix.
            1. I don't have too much opinion on this. Just against how
 'hacky' it feels.
- Updated the `get session` to also return the session name if available
(for better humen-readability)
- [x] invoke
    - Works after the cli client-controller API updates.
- Added an example since it wasn't _too_ obvious to use 😅 (I wasn't sure
if `task` just meant a command/prompt or not at first)

## Interactive
_Note: Interactive mode requires a release version, so it may be hard to
test fixes here. I may just be able to find bugs through a released
`kagent` and try to resolve them. It _seems_ to use most of the same
features as non-interactive -- aside from `chat` -- so most fixes done
in non-interactive could be ported to the interactive mode._

```sh
// 1. build images & cli
make create-kind-cluster helm-install build-cli-local

// 2. ensure kagent image uses the localhost:5001 build
// I manually updated ./helm/kagent/values.yaml -> `registry: "localhost:5001"`

// 3. Use kagent CLI
KAGENT_HELM_REPO=./helm/ ./go/bin/kagent-local
```

Note: We may remove interactive mode, but currently keeping it for the
`chat` command. I tried importing it to non-interactive but hit issues
because it is (currently) implemented to work on an interactive shell.

- [x] chat
- Update terminology from "team(s)" to "agent(s)", similar to changes
done in [this PR](https://github.com/kagent-dev/kagent/pull/674).
  - Fixed session creation by passing required `agentRef`.
- Made more human-readable by using session name instead of session ID
in some instances (e.g. listing sessions)
  - Filter session list based on selected agent.
- Updated interactive mode to check for port-forwarding in the main
thread instead of goroutine.
- The con is that it stops kagent from opening until it's done retry and
failing.
- The pro is that it makes sense to not to be able to use the ishell
until the status of the port-forward is known (which could also be seen
as the status of kagent-controller).
- I didn't make a port-forward failure an `os.Exit(1)` because it's
possible to `un/install` kagent through the shell, so the
port-forwarding state can be running/failing and it is valid.
- Updated interactive shell to store port-forward state (existent vs
non-existent) in its context and update it accordingly (i.e. on
`install`/`uninstall`).
- After <at>antweiss brought up port-forwarding issues, I noticed some
more quirks with the previous port-forwarding logic which led to all
these changes in order to have a cleaner UX.

## Misc

- Updated references from kagent @ port 8080 to kagent-controller @ port
8083.
- Updated port-forwarding logic to...
  1. rely on existing `StartPortForward` method 
  2. retry up to 1s, instead of 250ms

Resolves: https://github.com/kagent-dev/kagent/issues/659

---------

Signed-off-by: Fabian Gonzalez <fabian.gonzalez@solo.io>
Signed-off-by: Eitan Yarmush <eitan.yarmush@solo.io>
Co-authored-by: Eitan Yarmush <eitan.yarmush@solo.io>