feat: initial auth framework (#761)
lay the ground work for auth & authz in kagent. Initial impl leaves
current state - user is by default `admin@kagent.dev` and it's
authorized to do anything.

Added test agent - it replies with a constant text without using LLM.
This approach has these advantages
- It's fast & reliable (no call to slow LLM on the ineternet)
- It's predictable (we can reliable make test asserts)
- Low cost (no need for API key)
Which makes it ideal for integration / e2e tests for most of kagent
(with the exception of the bundled agents themselves).

---------

Signed-off-by: Yuval Kohavi <yuval.kohavi@gmail.com>
Signed-off-by: Eitan Yarmush <eitan.yarmush@solo.io>
Co-authored-by: Eitan Yarmush <eitan.yarmush@solo.io>