(Fix) Chat History Deletion (#857)
This PR fixes chat history not being able to be deleted and closes #627.
It updates the database client to use the "id" key instead of "name"
when deleting a session because "id" is the proper primary key of the
sessions table and is passed by the frontend and the handler.

Signed-off-by: Jet Chiang <jetjiang.ez@gmail.com>