feat: Set system message from ConfigMap or Secrets (#894)
Implements #625 

This PR allows user to set system message for declarative agents by
providing a `systemMessageFrom` field that specify either a `ConfigMap`
or `Secret` where the system message is stored, as an alternative to
specifying a string as `systemMessage` originally.

A minor refactoring was done on the types and the controller's RBAC
config is extended to read from config maps and secrets.

The usage is shown as below. I will update the documentation on
`kagent/website` once this is approved and merged.

1. Create configmap or secret:

```yaml
apiVersion: v1
kind: ConfigMap
metadata:
  name: my-agent-config
  namespace: kagent
data:
  system-message: "You will speak in the style of Shakespeare"
```


2. Create agent

```yaml
apiVersion: kagent.dev/v1alpha2
kind: Agent
metadata:
  name: agent-from-configmap
  namespace: kagent
spec:
  description: This agent speaks in Shakespearean dialect
  type: Declarative

  declarative:
    modelConfig: default-model-config
    systemMessageFrom:
      type: ConfigMap
      name: my-agent-config
      key: system-message
```

---------

Signed-off-by: Jet Chiang <jetjiang.ez@gmail.com>