[Feature] Implement Persistence and Tracing for CrewAI BYO Agents (#951)
This PR is a follow up for #920 

## Features

- [x] Creates go handlers for `/crewai` routes for storing and
retrieving memory items
- [x] Updates the database handler to store and retrieve long term
memory items from CrewAI Crew and state for CrewAI Flow
- [x] Adds custom memory store to agents before execution to allow for
session-based persistence
- [x] Updates samples to show usage of memory by setting `memory=True`
and persistence by `@persist()`
- [x] Support tracing with `opentelemetry-instrumentation-crewai` for
crewai specific spans as suggested in code review

## Tests

e2e test is added for CrewAI Poem flow sample agents using the mock LLM
server. The test case will create the agent resource, create the mock
LLM server using mock response in `invoke_creawi_agent.json`, test
synchronous, streaming, and persistence for the agent. It requires the
agent container to be build and pushed to the registry (by running `make
poem-flow-sample` or the Dockerfile directly in
`samples/crewai/poem_flow`).

The following changes are made to helper functions in e2e test:

1. `runSyncTest` accepts an optional `contextID` to be included in the
mock message to test session persistence
2. `runSyncTest` accepts an optional `useArtifacts` argument to indicate
if the expected output should be checked for in the history messages or
the artifact returned by the A2A server, since the A2A protocol
specifies that `Artifacts` are the standard way to convey final outputs
of a task

---------

Signed-off-by: Jet Chiang <jetjiang.ez@gmail.com>